# docs/architecture/specs/platform-api.openapi.yaml
openapi: 3.0.3
info:
  title: Trade Nexus Platform API
  version: 1.0.0
  description: |
    Canonical public API contract for Trade Nexus.

    Rules:
    - This file is the source of truth for external consumers.
    - Lona and execution engines are internal provider integrations behind adapters.
servers:
  - url: https://api.trade-nexus.io
    description: Production
  - url: http://localhost:3000
    description: Local development
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /v1/research/market-scan:
    post:
      tags: [Research]
      summary: Analyze market conditions and return strategy ideas
      operationId: postMarketScanV1
      parameters:
        - $ref: '#/components/parameters/XRequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarketScanRequest'
            example:
              assetClasses: [crypto, stocks]
              capital: 25000
              constraints:
                maxPositionPct: 25
                maxDrawdownPct: 12
      responses:
        '200':
          description: Research result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketScanResponse'
              example:
                requestId: req-research-001
                regimeSummary: Risk-on momentum with elevated volatility clusters.
                strategyIdeas:
                  - name: BTC Trend Follow
                    assetClass: crypto
                    description: Long breakouts with trailing stop.
                    rationale: Momentum persists while breadth remains positive.
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '429':
          $ref: '#/components/responses/Error429'
  /v1/strategies:
    get:
      tags: [Strategies]
      summary: List strategies
      operationId: listStrategiesV1
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/StrategyStatus'
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Strategy list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyListResponse'
              example:
                requestId: req-strategy-list-001
                items:
                  - id: strat-001
                    name: BTC Trend Follow
                    description: Breakout strategy for BTCUSD.
                    status: tested
                    provider: lona
                    providerRefId: lona-strategy-123
                    tags: [momentum, crypto]
                    createdAt: '2026-02-10T10:00:00Z'
                    updatedAt: '2026-02-13T19:00:00Z'
                nextCursor: null
        '401':
          $ref: '#/components/responses/Error401'
    post:
      tags: [Strategies]
      summary: Create strategy from natural-language description
      operationId: createStrategyV1
      parameters:
        - $ref: '#/components/parameters/XRequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStrategyRequest'
            example:
              name: Mean Reversion ETH
              description: Mean-reversion strategy for ETH using 20-period z-score bands.
              provider: xai
      responses:
        '201':
          description: Strategy created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyResponse'
              example:
                requestId: req-strategy-create-001
                strategy:
                  id: strat-002
                  name: Mean Reversion ETH
                  description: Mean-reversion strategy for ETH using z-score bands.
                  status: draft
                  provider: lona
                  providerRefId: lona-strategy-456
                  tags: [mean-reversion, crypto]
                  createdAt: '2026-02-13T21:20:00Z'
                  updatedAt: '2026-02-13T21:20:00Z'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
  /v1/strategies/{strategyId}:
    get:
      tags: [Strategies]
      summary: Get strategy by id
      operationId: getStrategyV1
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/StrategyId'
      responses:
        '200':
          description: Strategy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyResponse'
              example:
                requestId: req-strategy-get-001
                strategy:
                  id: strat-001
                  name: BTC Trend Follow
                  description: Breakout strategy for BTCUSD.
                  status: tested
                  provider: lona
                  providerRefId: lona-strategy-123
                  tags: [momentum, crypto]
                  createdAt: '2026-02-10T10:00:00Z'
                  updatedAt: '2026-02-13T19:00:00Z'
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
    patch:
      tags: [Strategies]
      summary: Update strategy metadata or status
      operationId: updateStrategyV1
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/StrategyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStrategyRequest'
            example:
              status: deployable
              tags: [momentum, approved]
      responses:
        '200':
          description: Updated strategy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyResponse'
              example:
                requestId: req-strategy-update-001
                strategy:
                  id: strat-001
                  name: BTC Trend Follow
                  description: Breakout strategy for BTCUSD.
                  status: deployable
                  provider: lona
                  providerRefId: lona-strategy-123
                  tags: [momentum, approved]
                  createdAt: '2026-02-10T10:00:00Z'
                  updatedAt: '2026-02-13T21:00:00Z'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
  /v1/strategies/{strategyId}/backtests:
    post:
      tags: [Backtests]
      summary: Start a backtest for a strategy
      operationId: createBacktestV1
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/StrategyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBacktestRequest'
            example:
              dataIds: [dataset-btc-1h-2025]
              startDate: '2025-01-01'
              endDate: '2025-12-31'
              initialCash: 100000
      responses:
        '202':
          description: Backtest accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BacktestResponse'
              example:
                requestId: req-backtest-create-001
                backtest:
                  id: bt-001
                  strategyId: strat-001
                  status: queued
                  startedAt: null
                  completedAt: null
                  metrics: {}
                  error: null
                  createdAt: '2026-02-13T21:25:00Z'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
  /v1/backtests/{backtestId}:
    get:
      tags: [Backtests]
      summary: Get backtest status and report
      operationId: getBacktestV1
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/BacktestId'
      responses:
        '200':
          description: Backtest report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BacktestResponse'
              example:
                requestId: req-backtest-get-001
                backtest:
                  id: bt-001
                  strategyId: strat-001
                  status: completed
                  startedAt: '2026-02-13T21:25:30Z'
                  completedAt: '2026-02-13T21:27:00Z'
                  metrics:
                    sharpeRatio: 1.48
                    maxDrawdownPct: 9.2
                    winRatePct: 57.3
                    totalReturnPct: 24.5
                  error: null
                  createdAt: '2026-02-13T21:25:00Z'
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
  /v1/deployments:
    get:
      tags: [Deployments]
      summary: List deployments
      operationId: listDeploymentsV1
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/DeploymentStatus'
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Deployment list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentListResponse'
              example:
                requestId: req-deployment-list-001
                items:
                  - id: dep-001
                    strategyId: strat-001
                    mode: paper
                    status: running
                    capital: 20000
                    engine: live-engine
                    providerRefId: live-abc
                    latestPnl: 150.25
                    createdAt: '2026-02-13T20:00:00Z'
                    updatedAt: '2026-02-13T21:10:00Z'
                nextCursor: null
        '401':
          $ref: '#/components/responses/Error401'
    post:
      tags: [Deployments]
      summary: Create a paper/live deployment
      operationId: createDeploymentV1
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeploymentRequest'
            example:
              strategyId: strat-001
              mode: paper
              capital: 20000
      responses:
        '202':
          description: Deployment accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentResponse'
              example:
                requestId: req-deployment-create-001
                deployment:
                  id: dep-001
                  strategyId: strat-001
                  mode: paper
                  status: queued
                  capital: 20000
                  engine: live-engine
                  providerRefId: live-abc
                  latestPnl: null
                  createdAt: '2026-02-13T21:30:00Z'
                  updatedAt: '2026-02-13T21:30:00Z'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '409':
          $ref: '#/components/responses/Error409'
  /v1/deployments/{deploymentId}:
    get:
      tags: [Deployments]
      summary: Get deployment status
      operationId: getDeploymentV1
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/DeploymentId'
      responses:
        '200':
          description: Deployment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentResponse'
              example:
                requestId: req-deployment-get-001
                deployment:
                  id: dep-001
                  strategyId: strat-001
                  mode: paper
                  status: running
                  capital: 20000
                  engine: live-engine
                  providerRefId: live-abc
                  latestPnl: 150.25
                  createdAt: '2026-02-13T20:00:00Z'
                  updatedAt: '2026-02-13T21:10:00Z'
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
  /v1/deployments/{deploymentId}/actions/stop:
    post:
      tags: [Deployments]
      summary: Stop a running deployment
      operationId: stopDeploymentV1
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/DeploymentId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
            example:
              reason: Risk policy drawdown threshold breached.
      responses:
        '202':
          description: Stop accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentResponse'
              example:
                requestId: req-deployment-stop-001
                deployment:
                  id: dep-001
                  strategyId: strat-001
                  mode: paper
                  status: stopping
                  capital: 20000
                  engine: live-engine
                  providerRefId: live-abc
                  latestPnl: 120.0
                  createdAt: '2026-02-13T20:00:00Z'
                  updatedAt: '2026-02-13T21:15:00Z'
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
  /v1/portfolios:
    get:
      tags: [Portfolios]
      summary: List portfolios
      operationId: listPortfoliosV1
      parameters:
        - $ref: '#/components/parameters/XRequestId'
      responses:
        '200':
          description: Portfolio list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioListResponse'
              example:
                requestId: req-portfolio-list-001
                items:
                  - id: portfolio-paper-001
                    mode: paper
                    cash: 12450.75
                    totalValue: 20891.12
                    pnlTotal: 891.12
                    positions:
                      - symbol: BTCUSDT
                        quantity: 0.3
                        avgPrice: 62000
                        currentPrice: 64800
                        unrealizedPnl: 840
        '401':
          $ref: '#/components/responses/Error401'
  /v1/portfolios/{portfolioId}:
    get:
      tags: [Portfolios]
      summary: Get portfolio snapshot
      operationId: getPortfolioV1
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/PortfolioId'
      responses:
        '200':
          description: Portfolio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioResponse'
              example:
                requestId: req-portfolio-get-001
                portfolio:
                  id: portfolio-paper-001
                  mode: paper
                  cash: 12450.75
                  totalValue: 20891.12
                  pnlTotal: 891.12
                  positions:
                    - symbol: BTCUSDT
                      quantity: 0.3
                      avgPrice: 62000
                      currentPrice: 64800
                      unrealizedPnl: 840
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
  /v1/orders:
    get:
      tags: [Orders]
      summary: List orders
      operationId: listOrdersV1
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
              example:
                requestId: req-order-list-001
                items:
                  - id: ord-001
                    symbol: BTCUSDT
                    side: buy
                    type: limit
                    quantity: 0.1
                    price: 64500
                    status: pending
                    deploymentId: dep-001
                    createdAt: '2026-02-13T21:31:00Z'
                nextCursor: null
        '401':
          $ref: '#/components/responses/Error401'
    post:
      tags: [Orders]
      summary: Place manual order
      operationId: createOrderV1
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              symbol: BTCUSDT
              side: buy
              type: limit
              quantity: 0.1
              price: 64500
              deploymentId: dep-001
      responses:
        '201':
          description: Order placed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
              example:
                requestId: req-order-create-001
                order:
                  id: ord-001
                  symbol: BTCUSDT
                  side: buy
                  type: limit
                  quantity: 0.1
                  price: 64500
                  status: pending
                  deploymentId: dep-001
                  createdAt: '2026-02-13T21:31:00Z'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '409':
          $ref: '#/components/responses/Error409'
  /v1/orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get order
      operationId: getOrderV1
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
              example:
                requestId: req-order-get-001
                order:
                  id: ord-001
                  symbol: BTCUSDT
                  side: buy
                  type: limit
                  quantity: 0.1
                  price: 64500
                  status: pending
                  deploymentId: dep-001
                  createdAt: '2026-02-13T21:31:00Z'
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
    delete:
      tags: [Orders]
      summary: Cancel order
      operationId: cancelOrderV1
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
              example:
                requestId: req-order-cancel-001
                order:
                  id: ord-001
                  symbol: BTCUSDT
                  side: buy
                  type: limit
                  quantity: 0.1
                  price: 64500
                  status: cancelled
                  deploymentId: dep-001
                  createdAt: '2026-02-13T21:31:00Z'
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
  /v1/datasets/uploads:init:
    post:
      tags: [Datasets]
      summary: Initialize dataset upload session
      operationId: initDatasetUploadV1
      parameters:
        - $ref: '#/components/parameters/XRequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasetUploadInitRequest'
            example:
              filename: btc-1h-2025.csv
              contentType: text/csv
              sizeBytes: 2048
      responses:
        '202':
          description: Dataset upload initialized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetUploadInitResponse'
              example:
                requestId: req-dataset-init-001
                datasetId: dataset-001
                uploadUrl: https://uploads.trade-nexus.local/dataset-001
                status: uploading
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    XRequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
      description: Caller-provided request id for trace correlation.
...
  responses:
    Error400:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    MarketScanRequest:
      type: object
      required: [assetClasses, capital]
      properties:
        assetClasses:
          type: array
          minItems: 1
          items:
            type: string
            enum: [crypto, stocks, forex, etf, commodities]
        capital:
          type: number
          minimum: 0
        constraints:
          type: object
          properties:
            maxPositionPct:
              type: number
              minimum: 0
              maximum: 100
            maxDrawdownPct:
              type: number
              minimum: 0
              maximum: 100
    MarketScanResponse:
      type: object
      required: [requestId, regimeSummary, strategyIdeas]
      properties:
        requestId:
          type: string
        regimeSummary:
          type: string
        strategyIdeas:
          type: array
          items:
            type: object
            required: [name, assetClass, description]
            properties:
              name:
                type: string
              assetClass:
                type: string
              description:
                type: string
              rationale:
                type: string
...